+++ contracts/mocks/token/ERC20/ERC20MetaMock.sol
@@ -80,22 +80,10 @@ contract ERC20MetaMock is ERC20Fees, ERC20 {
             maxPossibleCharge);
     }
 
-    /**
-     * @dev Replacement for msg.sender. Returns the actual sender of a transaction: msg.sender for regular transactions,
-     * and the end-user for GSN relayed calls (where msg.sender is actually `RelayHub`).
-     *
-     * IMPORTANT: Contracts derived from {GSNRecipient} should never use `msg.sender`, and use {_msgSender} instead.
-     */
     function _msgSender() internal override(Context, ERC20Fees) view returns (address payable) {
         return super._msgSender();
     }
 
-    /**
-     * @dev Replacement for msg.data. Returns the actual calldata of a transaction: msg.data for regular transactions,
-     * and a reduced version for GSN relayed calls (where msg.data contains additional information).
-     *
-     * IMPORTANT: Contracts derived from {GSNRecipient} should never use `msg.data`, and use {_msgData} instead.
-     */
     function _msgData() internal override(Context, ERC20Fees) view returns (bytes memory) {
         return super._msgData();
     }